+++ R/predict.ridgeLinear.R
@@ -1,5 +1,8 @@
 ## predict method for linear ridge regression models
 
+#' @rdname predict
+#' @export
+#' @importFrom stats na.pass terms model.matrix model.frame delete.response .checkMFClasses coef
 predict.ridgeLinear <- function(object, newdata,  
     na.action = na.pass, all.coef = FALSE, ...)
   {
@@ -7,7 +10,9 @@ predict.ridgeLinear <- function(object, newdata,
     if (!inherits(object, "ridgeLinear")) 
       warning("calling predict.ridgeLinear(<fake-ridgeLinear-object>) ...")
     if (missing(newdata) || is.null(newdata)) {
-      mm <- X <- model.frame(object)
+      # model.matrix handles factors properly, and includes an intercept term
+      # TODO(dan): Add xlev here?
+      mm <- X <- model.matrix(object, data=model.frame(formula=tt, data=object$model_frame))
       mmDone <- TRUE
       offset <- object$offset
     }
@@ -17,7 +22,7 @@ predict.ridgeLinear <- function(object, newdata,
                        xlev = object$xlevels)
       if (!is.null(cl <- attr(Terms, "dataClasses"))) 
         .checkMFClasses(cl, m)
-      mm <- X <- model.matrix(Terms, m)
+      mm <- X <- model.matrix(Terms, m, contrasts.arg = object$contrasts)
       offset <- rep(0, nrow(X))
       if (!is.null(off.num <- attr(tt, "offset"))) 
         for (i in off.num) offset <- offset + eval(attr(tt, 
@@ -26,12 +31,6 @@ predict.ridgeLinear <- function(object, newdata,
         offset <- offset + eval(object$call$offset, newdata)
       mmDone <- FALSE
     }
-    hasintercept <- attr(tt, "intercept")
-    ll <- attr(tt, "term.labels")
-    if(hasintercept)
-      mm <- cbind(1, X[,ll,drop=FALSE])
-    else
-      mm <- X[,ll,drop=FALSE]
     beta <- coef(object, all.coef = all.coef)
     if(all.coef)
       res <- apply(beta, 1, function(x){drop(as.matrix(mm) %*% x)})